+++ client/Pienyritysappi/app/src/main/AndroidManifest.xml
@@ -18,10 +18,10 @@
         <activity android:name=".ReservationActivity" />
         <activity android:name=".LogIn" />
         <activity android:name=".JobInfo" />
-        <activity android:name=".JobOrder" />
+        <activity android:name=".CompanyServices" />
         <activity android:name=".Company" />
-        <activity android:name=".Category" />
-        <activity android:name=".Services" />
+        <activity android:name=".ChosenCategory" />
+        <activity android:name=".Categories" />
         <activity android:name=".Register" />
         <activity android:name=".MainActivity">
             <intent-filter>
