+++ client/Pienyritysappi/app/src/main/java/com/example/pienyritysappi/ChosenCategory.java
@@ -3,7 +3,6 @@ package com.example.pienyritysappi;
 import androidx.appcompat.app.AppCompatActivity;
 
 import android.content.Intent;
-import android.graphics.Color;
 import android.os.Bundle;
 import android.view.View;
 import android.widget.Button;
@@ -21,43 +20,46 @@ import org.json.JSONArray;
 import org.json.JSONException;
 import org.json.JSONObject;
 
-public class JobOrder extends AppCompatActivity {
+public class ChosenCategory extends AppCompatActivity {
 
-
+    private TextView mTextViewCategoryName;
     private RequestQueue mQueue;
-    private String url;
-    private String user_id = "";
+    private String categoryurl;
+    private int companyCount =1;
     private Button nButton;
+    private String companyName;
+    private String contactName;
     private JSONArray jsonArray;
-    private JSONObject job;
-    private int jobCount;
+    private JSONObject company;
+    private String baseUrl = "http://mobiilisovellus.therozor.com:5000/providers?category_id=";
+    private String categoryId;
+    private String apikey;
+
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_job_order);
+        setContentView(R.layout.activity_chosen_category);
         Intent intent = getIntent();
-        user_id = getIntent().getStringExtra("user_id");
-        url = "http://mobiilisovellus.therozor.com:5000/services?user_id="+user_id;
-        System.out.println("JobOrder url: "+ url);
+        categoryId = intent.getStringExtra("categoryId");
+        categoryurl = baseUrl + categoryId;
+        mTextViewCategoryName = findViewById(R.id.textViewCategoryName);
         mQueue = Volley.newRequestQueue(this);
         jsonParse();
-
     }
 
     private void jsonParse() {
-        JsonObjectRequest request = new JsonObjectRequest(Request.Method.GET, url, null,
+        JsonObjectRequest request = new JsonObjectRequest(Request.Method.GET, categoryurl, null,
                 new Response.Listener<JSONObject>() {
                     @Override
                     public void onResponse(JSONObject response) {
                         try {
                             jsonArray = response.getJSONArray("data");
-                            jobCount = response.getInt("count");
-
-                            for(int i = 0; i< jobCount; i++) {
+                            companyCount = response.getInt("count");
+                            mTextViewCategoryName.setText(response.getString("category_name"));
+                            for(int i = 0; i< companyCount; i++) {
                                 addButton(i); //luo myös onClickListenerin
                             }
-
                         } catch (JSONException e) {
                             System.out.println("\nnyt ollaan onResponsen catchissä: JSONExceptionissa\n");
                             e.printStackTrace();
@@ -74,30 +76,27 @@ public class JobOrder extends AppCompatActivity {
     }
 
     private void addButton(int i) {
-        GridLayout layout = (GridLayout)findViewById(R.id.jobButtonGridLayout);
+        GridLayout layout = (GridLayout)findViewById(R.id.CompanyButtonGridLayout);
         nButton = new Button(this);
         try {
-            job = jsonArray.getJSONObject(i);
-            String jobName = job.getString("service_title");
-            String jobAvailability = job.getString("service_availability");
-            String jobPrice = job.getString("service_price");
-            String jobId = job.getString("service_id");
-            String buttonText = jobName + "\nSaatavilla: " + jobAvailability + "\n" + jobPrice + " €";
+            company = jsonArray.getJSONObject(i);
+            companyName = company.getString("user_company_name");
+            contactName = company.getString("user_name");
+            String buttonText = companyName + "\nYhteyshenkilö: " + contactName;
+            String companyUserId = company.getString("user_id");
 
             GridLayout.LayoutParams params = new GridLayout.LayoutParams();
             params.setMargins(10,10,10,10);
-            nButton.setText(buttonText);
-            nButton.setTextAlignment(View.TEXT_ALIGNMENT_TEXT_START);
             nButton.setLayoutParams(params);
             nButton.setBackground(nButton.getContext().getDrawable(R.drawable.rounded_button));
+            nButton.setText(buttonText);
             nButton.setPadding(30,0,30,0);
-            String jobBaseUrl = "http://mobiilisovellus.therozor.com:5000/service/";
-            String joburl = jobBaseUrl + jobId;
+            nButton.setTextAlignment(View.TEXT_ALIGNMENT_TEXT_START);
             nButton.setOnClickListener(new View.OnClickListener(){
                 @Override
                 public void onClick(View v) {
-                    Intent intent = new Intent(getApplicationContext(),JobInfo.class);
-                    intent.putExtra("keyurl", joburl);
+                    Intent intent = new Intent(getApplicationContext(),Company.class);
+                    intent.putExtra("userId", companyUserId);
                     startActivity(intent);
                 }
             });
@@ -107,21 +106,13 @@ public class JobOrder extends AppCompatActivity {
         layout.addView(nButton);
     }
 
-
+    public void homeButtonClicked(View view) {
+        Intent intent = new Intent(getApplicationContext(),MainActivity.class);
+        startActivity(intent);
+    }
 
     public void profileButtonClicked(View view) {
-        Intent intent = new Intent(getApplicationContext(), CustomerProfile.class);
-        startActivity(intent);
-    }
-
-    public void homeButtonClicked(View view) {
-        Intent intent = new Intent(getApplicationContext(), MainActivity.class);
-        startActivity(intent);
-    }
-
-    public void buttonJobClicked(View view) {
-        Intent intent = new Intent(getApplicationContext(), JobInfo.class);
-
+        Intent intent = new Intent(getApplicationContext(),CustomerProfile.class);
         startActivity(intent);
     }
 }
