+++ client/Pienyritysappi/app/src/main/res/layout/activity_categories.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="#ADD8E6"
-    tools:context=".Services">
+    tools:context=".Categories">
 
     <TextView
         android:id="@+id/header"
